requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
boto3==1.28.53
pydantic==2.3.0
typer==0.9.0
rich==13.5.2
reportlab==4.0.4
jinja2==3.1.2
python-multipart==0.0.6
weasyprint==59.0
requests==2.31.0
python-dotenv==1.0.0
plotly==6.3.0
dash==3.2.0
pandas==2.3.2
